remove ProcesssingView
